::FEAT:: refatora o componente Cadastro e atualiza a navegação no Header
