+++ src/components/header.jsx
@@ -1,8 +1,9 @@
 import "./header.css";
 import { Container, Nav, Navbar } from "react-bootstrap";
 import { Link } from 'react-router-dom';
+import { handleOpenCadastro } from "./cadastro";
 
-function Header({ onOpenCadastro }) {
+function Header() {
   return (
     <>
       <Navbar bg="dark" variant="dark" expand="lg" className="justify-content-center">
@@ -21,7 +22,7 @@ function Header({ onOpenCadastro }) {
               <Nav.Link href="/">Home</Nav.Link>
               <Nav.Link as={Link} to="/indicacao">Grupos de Apoio</Nav.Link>
               <Nav.Link as={Link} to ="/login">Login</Nav.Link>
-              <Nav.Link href="#singIn" onClick={onOpenCadastro}>Cadastro</Nav.Link>
+              <Nav.Link onClick={handleOpenCadastro}>Cadastro</Nav.Link>
             </Nav>
           </Navbar.Collapse>
         </Container>
