+++ src/components/login.css
@@ -55,6 +55,7 @@ form input {
   border-radius: 8px;
   font-size: 16px;
   background: #f9f9f9;
+  color: #333;
   transition: border-color 0.3s;
 }
