+++ src/components/login.jsx
@@ -1,4 +1,5 @@
 import "./login.css";
+import { Link } from 'react-router-dom';
 
 function TelaDeLogin() {
   const realizarLogin = (e) => {
